implementing some safty

+++ bash_task1.sh
@@ -1,11 +1,16 @@
 #!/bin/env bash
 
+shopt -s nullglob
+
 dir="${1:?Usage: $0 <dir>}"
 
 find "$dir" -type f -name "*.mp4" -print0 | while IFS= read -d -r '' file; do
 
 	base="${file%.mp4}"
 	audio="${base}.m4a"
+
+	[[ -f "$audio" ]] || continue
+	
 	output="${base}.mux.mp4"
 	ffmpeg -i "$file" -i "$audio" -c copy "$output" && rm -- "$file" "$audio"
 	done
